Move test to IoTest
